+++ src/ansibleguy-webui/main.py
@@ -93,7 +93,7 @@ def main():
     django_setup()
     environ['AW_INIT'] = '0'
 
-    from db import create_first_superuser, create_manager_groups
+    from db import create_first_superuser, create_manager_groups, cleanup_job_stati
     from handle_signals import handle_signals
     from webserver import init_webserver
     from aw.execute.scheduler import init_scheduler
@@ -103,5 +103,6 @@ def main():
 
     create_first_superuser()
     create_manager_groups()
+    cleanup_job_stati()
     init_scheduler(handle_signals)
     init_webserver()