+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fix for Log-Directory creation
 * Enable SSH host-key checking by default (*ansible-runner seem to disable it by default*) 
 * Enhanced Bottons of Log Live-View
+* Git-Repository Cleanup-Hook (Post Job-Run)
 
 ----